GD/std string cleanup + separate data loading cpp
