+++ extension/src/register_types.h
