+++ game/src/Autoload/Events.gd
@@ -10,6 +10,7 @@ var _define_filepaths_dict : Dictionary = {
 	GameSingleton.get_water_province_file_key(): "res://common/map/water.json",
 	GameSingleton.get_region_file_key(): "res://common/map/regions.json",
 	GameSingleton.get_terrain_variant_file_key(): "res://common/map/terrain.json",
+	GameSingleton.get_terrain_texture_dir_key(): "res://art/terrain/",
 	GameSingleton.get_province_image_file_key(): "res://common/map/provinces.png",
 	GameSingleton.get_terrain_image_file_key(): "res://common/map/terrain.png",
 	GameSingleton.get_goods_file_key(): "res://common/goods.json",
